+++ z-paging/uni_modules/z-paging/components/z-paging/js/z-paging-utils.js
@@ -73,7 +73,7 @@ function getTouchFromZPaging(target) {
 			return getTouchFromZPaging(target.parentNode);
 		}
 	} else {
-		return {isFromZp: false};
+		return { isFromZp: false };
 	}
 }