+++ z-paging/uni_modules/z-paging/changelog.md
@@ -1,3 +1,5 @@
-## 2.5.3（2023-02-13）
-1.`新增` `reload`、`refresh`、`complete`、`completeByTotal`、`completeByNoMore`、`completeByKey`、`setLocalPaging`等方法的返回值添加请求结果promise，可以通过方法名().then()获取本次操作请求结束后的【总列表】和【是否有更多数据】。  
-2.`修复` 由`v2.5.0`引出的在vue2 + 钉钉小程序中，`slot="top"`插入无效的问题。  
+## 2.5.4（2023-03-05）
+1.`修复` `show-loading-more-when-reload`设置为true时，调用reload不显示`正在加载...`的问题。  
+2.`修复`  `isOldWebView`中的系统版本号解析错误导致部分系统中首次加载内容出现延迟的问题。  
+3.`优化` `refresher-enabled`被动态设置为false时，强制关闭当前下拉刷新。  
+4.`优化` 整合`z-paging`与`z-paging-swiper`共用代码并精简代码写法。